Add name check for <= v4 datasets

+++ split_compile.py
@@ -42,8 +42,11 @@ def main() -> None:
     # Loop all versions
     overall_cnt: int = 0
     for cv_idx, cv_ver in enumerate(const.CV_VERSIONS):
-        # Calc CV_DIR - TODO different for v1-4 !!!
-        cv_dir_name: str = "cv-corpus-" + cv_ver + "-" + const.CV_DATES[cv_idx]
+        # Calc CV_DIR - Different for v1-4 !!!
+        if cv_ver in ['1', '2', '3', '4']:
+            cv_dir_name: str = "cv-corpus-" + cv_ver
+        else:
+            cv_dir_name: str = "cv-corpus-" + cv_ver + "-" + const.CV_DATES[cv_idx]
         # Check if it exists in source (check "s1", if not there, it is nowhere)
         if not os.path.isdir(os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[0], cv_dir_name)):
             continue # Does, not exist, so skip
@@ -68,53 +71,66 @@ def main() -> None:
             overall_cnt += 1
             cnt += 1
             print('\033[F' + ' ' * 80)
-            print(f'\033[FProcessing locale {cnt}/{len(lc_list)} : {lc}')
+            print(f'\033[FProcessing version {cv_ver} ({cv_idx+1}/{len(const.CV_VERSIONS)}) locale {cnt}/{len(lc_list)} : {lc}')
 
             # copy splitting algorithm independent files
             src_dir: str = os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[0], cv_dir_name, lc)
             dst_dir: str = os.path.join(vc_dir_base, cv_dir_name, lc)
-            os.makedirs(os.path.join(dst_dir, const.ALGORITHMS[0]), exist_ok=True)
-            for fn in ["validated.tsv", "invalidated.tsv", "other.tsv", "reported.tsv"]:
-                if os.path.isfile(os.path.join(src_dir, fn)):
-                    shutil.copy2(os.path.join(src_dir, fn), dst_dir)
+            if conf.FORCE_CREATE or not os.path.isdir(dst_dir):
+                os.makedirs(os.path.join(dst_dir, const.ALGORITHMS[0]), exist_ok=True)
+                for fn in ["validated.tsv", "invalidated.tsv", "other.tsv", "reported.tsv"]:
+                    if os.path.isfile(os.path.join(src_dir, fn)):
+                        shutil.copy2(os.path.join(src_dir, fn), dst_dir)
 
             # copy to s1
             dst_dir: str = os.path.join(vc_dir_base, cv_dir_name, lc, const.ALGORITHMS[0])
-            for fn in ["train.tsv", "dev.tsv", "test.tsv"]:
-                if os.path.isfile(os.path.join(src_dir, fn)):
-                    shutil.copy2(os.path.join(src_dir, fn), dst_dir)
+            if conf.FORCE_CREATE or not os.path.isdir(dst_dir):
+                os.makedirs(dst_dir, exist_ok=True)
+                for fn in ["train.tsv", "dev.tsv", "test.tsv"]:
+                    if os.path.isfile(os.path.join(src_dir, fn)):
+                        shutil.copy2(os.path.join(src_dir, fn), dst_dir)
 
             # check if exists to copy to s99
             src_dir: str = os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[1], cv_dir_name, lc)
             if os.path.isdir(src_dir):
                 dst_dir: str = os.path.join(vc_dir_base, cv_dir_name, lc, const.ALGORITHMS[1])
-                os.makedirs(dst_dir, exist_ok=True)
-                for fn in ["train.tsv", "dev.tsv", "test.tsv"]:
-                    if os.path.isfile(os.path.join(src_dir, fn)):
-                        shutil.copy2(os.path.join(src_dir, fn), dst_dir)
+                if conf.FORCE_CREATE or not os.path.isdir(dst_dir):
+                    os.makedirs(dst_dir, exist_ok=True)
+                    for fn in ["train.tsv", "dev.tsv", "test.tsv"]:
+                        if os.path.isfile(os.path.join(src_dir, fn)):
+                            shutil.copy2(os.path.join(src_dir, fn), dst_dir)
 
             # check if exists to copy to v1
             src_dir: str = os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[2], cv_dir_name, lc)
             if os.path.isdir(src_dir):
                 dst_dir: str = os.path.join(vc_dir_base, cv_dir_name, lc, const.ALGORITHMS[2])
-                os.makedirs(dst_dir, exist_ok=True)
-                for fn in ["train.tsv", "dev.tsv", "test.tsv"]:
-                    if os.path.isfile(os.path.join(src_dir, fn)):
-                        shutil.copy2(os.path.join(src_dir, fn), dst_dir)
+                if conf.FORCE_CREATE or not os.path.isdir(dst_dir):
+                    os.makedirs(dst_dir, exist_ok=True)
+                    for fn in ["train.tsv", "dev.tsv", "test.tsv"]:
+                        if os.path.isfile(os.path.join(src_dir, fn)):
+                            shutil.copy2(os.path.join(src_dir, fn), dst_dir)
 
             # special case for clip durations, we need to find it
-            dst_dir: str = os.path.join(cd_dir_base, lc)
-            os.makedirs(dst_dir, exist_ok=True)
-            cd_file: str = '$clip_durations.tsv'
-            search_paths: list[str] = [
-                os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[0], cv_dir_name, lc, 'clips', cd_file),
-                os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[1], cv_dir_name, lc, 'clips', cd_file),
-                os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[2], cv_dir_name, lc, 'clips', cd_file)
-            ]
-            for p in search_paths:
-                if os.path.isfile(p):
-                    shutil.copy2(p, dst_dir)
-                    continue
+            # But dor this only for the latest version !!!
+            if cv_ver == const.CV_VERSIONS[-1]:
+                dst_dir: str = os.path.join(cd_dir_base, lc)
+                os.makedirs(dst_dir, exist_ok=True)
+                cd_file: str = '$clip_durations.tsv'
+                cd_found: bool = False
+                search_paths: list[str] = [
+                    os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[0], cv_dir_name, lc, 'clips', cd_file),
+                    os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[1], cv_dir_name, lc, 'clips', cd_file),
+                    os.path.join(conf.SRC_BASE_DIR, const.ALGORITHMS[2], cv_dir_name, lc, 'clips', cd_file)
+                ]
+                for p in search_paths:
+                    if os.path.isfile(p):
+                        shutil.copy2(p, dst_dir)
+                        cd_found = True
+                        continue
+                # TODO: If it is not found, we need to create it.
+                if not cd_found:
+                    print('\n\nWARNING: $clip_durations.tsv file not found for', lc, '\n')
+
 
     # done
     finish_time: datetime = datetime.now()
